refactor(mesi): split parser.go into config.go + parser.go
Move EsiParserConfig struct, all config methods, and discardLogger
to new config.go. Move config-related tests to config_test.go.
parser.go now focuses solely on orchestration (MESIParse, Parse,
assembleResults).

No functional changes. All tests pass, go doc diff is empty.
